# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the api-documentation of the QHighschool api",
    "version": "1.0.0",
    "title": "QHighschool api",
    "contact": {
      "email": "steven.bronsveld@gmail.com"
    }
  },
  "basePath": "/api",
...
  "paths": {
...
    "/user": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Get user data",
        "operationId": "getUser",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "formData",
            "description": "The userId of the user to get",
            "required": false,
            "type": "string"
          }
        ]
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Update user data",
        "operationId": "setUser",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "formData",
            "description": "The userId to change, if empty, own user will be returned",
            "required": false,
            "type": "string"
          },
          {
            "name": "preferedEmail",
            "in": "formData",
            "description": "The new prefered email",
            "required": true,
            "type": "string"
          },
          {
            "name": "profile",
            "in": "formData",
            "description": "The profile of the student",
            "required": true,
            "type": "string"
          },
          {
            "name": "phoneNumber",
            "in": "formData",
            "description": "Thephone number of the student",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
    "/user/enrollments": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get choices",
        "operationId": "getEnrollments",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "maxLength": 5,
              "items": {
                "$ref": "#/definitions/Group"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Add enrollment",
        "operationId": "addEnrollment",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "formData",
            "description": "The groupId of the group to add",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete enrollment",
        "operationId": "removeEnrollment",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "formData",
            "description": "The groupId of the group to remove",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "school": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "year": {
          "type": "integer",
          "format": "int64"
        },
        "level": {
          "type": "string",
          "enum": [
            "vwo",
            "havo",
            "mavo"
          ]
        },
        "preferedEmail": {
          "type": "string"
        },
        "profile": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Course"
      }
    },
    "Group": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "teacherId": {
          "type": "string"
        },
        "teacherName": {
          "type": "string"
        },
        "day": {
          "type": "string",
          "enum": [
            "maandag",
            "dinsdag",
            "woensdag",
            "donderdag",
            "vrijdag"
          ]
        },
        "schoolYear": {
          "type": "string",
          "enum": [
            "2018/2019",
            "2019/2020",
            "2021/2022",
            "2023/2024"
          ]
        },
        "period": {
          "type": "integer"
        },
        "courseId": {
          "type": "integer"
        },
        "courseName": {
          "type": "string"
        },
        "courseDescription": {
          "type": "string"
        },
        "subjectId": {
          "type": "integer"
        },
        "subjectName": {
          "type": "string"
        },
        "subjectDescription": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Group"
      }
    },
...
  }
}